docker/requirements.txt
asgiref==3.8.1
Babel==2.15.0
Django==5.0.6
django-cors-headers==4.3.1
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
phonenumberslite==8.13.36
python-dotenv==1.0.1
sqlparse==0.5.0
